requirements.txt
fastapi==0.78.0
Jinja2==2.10
RPi.GPIO==0.7.0
uvicorn==0.17.6
starlette==0.19.1